+++ common.php
@@ -1,10 +1,13 @@
 <?php
 
+// don't show warnings in BBCodeParser2
+$old_error_reporting = error_reporting(E_ERROR);
 include('HTML/BBCodeParser2.php');
 $bbopt = array(
 	'filters' => array('Basic', 'Email', 'Extended', 'Images', 'Links', 'Lists')
 );
 $bb = new HTML_BBCodeParser2($bbopt);
+error_reporting($old_error_reporting);
 
 function template_print($var, $template) {
 	extract($var);
@@ -24,6 +27,13 @@ function template_get($var, $template) {
 	return $res;
 }
 
+function mkdir_p($dir) {
+	if (!is_dir($dir) && !is_link($dir)) {
+		mkdir($dir, 0755);
+	}
+}
+
+
 function write_content($file, $content) {
 	global $target_dir;
 
@@ -31,13 +41,8 @@ function write_content($file, $content) {
 	$dir = dirname($target);
 
 	$one_up_dir = dirname($dir);
-	if (!is_dir($one_up_dir) && !is_link($one_up_dir)) {
-		mkdir($one_up_dir, 0755);
-	}
-
-	if (!is_dir($dir) && !is_link($dir)) {
-		mkdir($dir, 0755);
-	}
+	mkdir_p($one_up_dir);
+	mkdir_p($dir);
 
 	$f = fopen($target, 'w');
 	fputs($f, $content);
@@ -58,3 +63,38 @@ function slug($str) {
 		$str = trim($str, '-');
 		return $str;
 }
+
+function get_topic_title_with_type($topic_title, $topic_type) {
+  global $topic_title_prefix_sticky, $topic_title_prefix_announcement, $topic_title_prefix_global;
+
+	if ($topic_type == 1) {
+		// sticky
+		return $topic_title_prefix_sticky . $topic_title;
+	} else if ($topic_type == 2) {
+		// announcement
+		return $topic_title_prefix_announcement . $topic_title;
+	} else if ($topic_type == 3) {
+		// global
+		return $topic_title_prefix_global . $topic_title;
+	}
+	return $topic_title;
+}
+
+
+// from: https://stackoverflow.com/questions/2050859/copy-entire-contents-of-a-directory-to-another-using-php
+function recurse_copy($src, $dst) { 
+    $dir = opendir($src); 
+    mkdir_p($dst);
+    while(false !== ( $file = readdir($dir)) ) { 
+        if (( $file != '.' ) && ( $file != '..' )) { 
+            if ( is_dir($src . '/' . $file) ) { 
+                recurse_copy($src . '/' . $file,$dst . '/' . $file); 
+            } 
+            else { 
+                copy($src . '/' . $file,$dst . '/' . $file); 
+            } 
+        } 
+    } 
+    closedir($dir); 
+} 
+